backend/ocr_platform/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
black==25.12.0
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distlib==0.4.0
fastapi==0.128.0
filelock==3.20.1
flake8==7.3.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiwer==4.0.0
librt==0.7.5
MarkupSafe==3.0.3
mccabe==0.7.0
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.7
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.21
pytokens==0.3.0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
virtualenv==20.35.4
